updated test class

+++ src/main/java/com/cg/fds/service/CustomerServiceImp.java
@@ -8,14 +8,20 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import com.cg.fds.entities.Address;
 import com.cg.fds.entities.Customer;
+import com.cg.fds.entities.FoodCart;
+import com.cg.fds.exception.CustomerNotFoundException;
 import com.cg.fds.exception.InvalidCustomerAddressException;
 import com.cg.fds.exception.InvalidCustomerException;
 import com.cg.fds.exception.InvalidCustomerPhoneNumberException;
+import com.cg.fds.exception.RemoveCustomerException;
+import com.cg.fds.exception.UpdateCustomerException;
 import com.cg.fds.repository.ICustomerRepository;
 
 public class CustomerServiceImp implements ICustomerService {
 	@Autowired
 	ICustomerRepository customerRepository;
+	
+	
 
 	public String generateId() {
 		StringBuilder builder= new StringBuilder();
@@ -30,21 +36,27 @@ public class CustomerServiceImp implements ICustomerService {
 	@Override
 	public Customer addCustomer(Customer customer) {
 		validateCustomer(customer);
-		validatePhone(customer.getMobileNumber());
-		Address address=customer.getAddress();
-		Address savedAddress=addressRepository.save(address);
-		customer.setAddress(savedAddress);
-		customer.setCustomerId(id);
+//		validatePhone(customer.getMobileNumber());
+//		Address address=customer.getAddress();
+//		Address savedAddress=addressRepository.save(address);
+//		customer.setAddress(savedAddress);
+//		customer.setCustomerId(customer.id);
 		Customer saved = customerRepository.save(customer);
-		FoodCart cart=new FoodCart();
-		cart.setCustomer(saved);
-		cartRepository.save(cart);
+//		FoodCart cart=new FoodCart();
+//		cart.setCustomer(saved);
+		//cartRepository.save(cart);
 		return saved;
 	}
 
 	@Override
 	public Customer updateCustomer(Customer customer) {
 		validateCustomer(customer);
+		String customerId = customer.getCustomerId();
+		boolean exist=customerRepository.existsById("1");
+		if(!exist)
+		{
+			throw new UpdateCustomerException("Customer doesn't exist for id =" + customer.getCustomerId());
+		}
 		Customer updateCustomer = customerRepository.save(customer);
 
 		return updateCustomer;
@@ -53,18 +65,23 @@ public class CustomerServiceImp implements ICustomerService {
 	@Override
 	public Customer removeCustomer(Customer customer) {
 		validateCustomer(customer);
+		String customerId = customer.getCustomerId();
+		boolean exist=customerRepository.existsById("1");
+		if(!exist)
+		{
+			throw new RemoveCustomerException("Customer doesn't exist for id =" + customer.getCustomerId());
+		}
 		Customer removeCustomer = customerRepository.remove(customer);
 		return removeCustomer;
 	}
 
 	@Override
-	public Customer viewCustomer(int id) {
+	public Customer viewCustomer(String id) {
 		Optional<Customer> viewCustomer = customerRepository.findById(id);
-		Customer customer = null;
-		if (viewCustomer.isPresent()) {
-			customer = viewCustomer.get();
+		if (!viewCustomer.isPresent()) {
+			throw new CustomerNotFoundException("Customer doesn't exist for id =" + id);
 		}
-		return customer;
+		return viewCustomer.get();
 	}
 
 	@Override
